Move to GHC 9.2.8 package-set
Moves to using GHC 9.2.8 as the base package-set,
updating from the earlier GHC 9.2.7 package set.

GHC 9.2.8 is the compiler version that's best supported
by HLS in the 9.2 series, looking at it's support table.

This would bring the best-case scenario for HLS for us,
as much as we can hope on 9.2 series.